+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 ## Change log
 ----------------------
 
+Version 8.1-SNAPSHOT
+-------------
+
+ADDED:
+
+- new method in the SecureRandomBuilder with the seed argument as Date object
+
+CHANGED:
+
+- update of gradle version to 6.8.3
+- update of silly-collections version to 8.4
+- update of test dependency equalsverifier version to 3.5.5
+- to new package io.github.astrapi69
+
 Version 8
 -------------